requirements.txt
streamlit
requests
pdfplumber
sentence-transformers
numpy
google-generativeai
torch
